feedback + tests
