+++ openvmm/openvmm_entry/src/ttrpc/mod.rs
@@ -10,14 +10,19 @@
 mod fd_passing;
 
 #[cfg(unix)]
-use fd_passing::FdRegistry;
+mod unix;
+#[cfg(windows)]
+mod windows;
 
-/// On non-unix platforms the fd-passing protocol does not exist. The registry
-/// is still threaded through the shared NIC configuration code, so provide an
-/// empty placeholder there; it is never populated or resolved.
-#[cfg(not(unix))]
-#[derive(Clone, Default)]
-struct FdRegistry {}
+#[cfg(unix)]
+use unix as platform;
+#[cfg(windows)]
+use windows as platform;
+
+use platform::BoundListener;
+use platform::FdRegistry;
+pub use platform::Listener;
+pub use platform::listener_from_path;
 
 use crate::cli_args::GuestPowerAction;
 use crate::meshworker::VmmMesh;
@@ -159,37 +164,9 @@ pub struct Parameters {
     pub transport: RpcTransport,
 }
 
-#[derive(mesh::MeshPayload)]
-pub enum Listener {
-    Unix(UnixListener),
-    #[cfg(windows)]
-    Pipe {
-        path: String,
-        allow_sid: Option<String>,
-    },
-}
-
-pub fn listener_from_path(
-    path: &std::path::Path,
-    allow_sid: Option<&str>,
-) -> anyhow::Result<Listener> {
-    #[cfg(windows)]
-    if mesh_rpc::is_named_pipe_path(path) {
-        return Ok(Listener::Pipe {
-            path: path.to_string_lossy().into_owned(),
-            allow_sid: allow_sid.map(str::to_owned),
-        });
-    }
-
-    #[cfg(not(windows))]
-    anyhow::ensure!(
-        allow_sid.is_none(),
-        "allow_sid is only supported on Windows"
-    );
+fn bind_unix_listener(path: &std::path::Path) -> anyhow::Result<UnixListener> {
     let _ = std::fs::remove_file(path);
-    Ok(Listener::Unix(
-        UnixListener::bind(path).context("failed to bind to socket")?,
-    ))
+    UnixListener::bind(path).context("failed to bind to socket")
 }
 
 #[derive(Copy, Clone, mesh::MeshPayload)]
@@ -266,8 +243,6 @@ mod dispatch {
     use super::FdRegistry;
     use super::ResolvedTransport;
     use futures::AsyncRead;
-    #[cfg(windows)]
-    use futures::AsyncReadExt;
     use futures::AsyncWrite;
     use futures::FutureExt;
     use pal_async::driver::Driver;
@@ -315,55 +290,26 @@ mod dispatch {
         Ok(())
     }
 
-    #[cfg(windows)]
-    pub(super) async fn run_pipe(
-        server: &mesh_rpc::Server,
-        driver: &(impl Driver + ?Sized),
-        listener: pal_async::windows::pipe::NamedPipeServer,
-        cancel: mesh::OneshotReceiver<()>,
-        transport: ResolvedTransport,
-    ) -> anyhow::Result<()> {
-        let mut tasks = FuturesUnordered::new();
-        let mut cancel = cancel.fuse();
-        let mut accept = std::pin::pin!(listener.accept(driver)?.fuse());
-        loop {
-            futures::select! { // merge semantics
-                result = accept => {
-                    accept.set(listener.accept(driver)?.fuse());
-                    if let Ok(conn) = result.and_then(|conn| pal_async::pipe::PolledPipe::new(driver, conn)) {
-                        tasks.push(async move {
-                            let _ = serve_pipe(server, conn, transport)
-                                .await
-                                .map_err(|err| {
-                                    tracing::error!(
-                                        error = err.as_ref() as &dyn std::error::Error,
-                                        "connection error"
-                                    )
-                                });
-                        });
-                    }
-                }
-                _ = tasks.next() => continue,
-                _ = cancel => break,
-            }
-        }
-        Ok(())
-    }
-
-    async fn serve_by_protocol<T>(
+    pub(super) async fn serve_by_protocol<T>(
         server: &mesh_rpc::Server,
         conn: T,
         first_byte: u8,
         transport: ResolvedTransport,
+        shutdown_policy: mesh_rpc::server::ShutdownPolicy,
     ) -> anyhow::Result<()>
     where
         T: AsyncRead + AsyncWrite + Unpin,
     {
+        #[cfg(not(feature = "grpc"))]
+        let _ = shutdown_policy;
+
         match first_byte {
             #[cfg(feature = "ttrpc")]
             0x00 if transport.allows_ttrpc() => server.serve_connection(conn).await,
             #[cfg(feature = "grpc")]
-            b'P' if transport.allows_grpc() => server.serve_connection_grpc(conn).await,
+            b'P' if transport.allows_grpc() => {
+                server.serve_connection_grpc(conn, shutdown_policy).await
+            }
             byte => {
                 anyhow::bail!("unrecognized or disallowed rpc protocol (first byte {byte:#04x})")
             }
@@ -387,84 +333,7 @@ mod dispatch {
             return Ok(());
         };
 
-        // The fd-passing protocol is allowed in every transport mode; it
-        // shares the socket with ttrpc/gRPC and is selected by its distinct
-        // magic first byte.
-        #[cfg(unix)]
-        if first_byte == super::fd_passing::MAGIC_FIRST_BYTE {
-            return super::fd_passing::serve(conn, registry).await;
-        }
-
-        #[cfg(not(unix))]
-        let _ = registry;
-
-        serve_by_protocol(server, conn, first_byte, transport).await
-    }
-
-    #[cfg(windows)]
-    async fn serve_pipe(
-        server: &mesh_rpc::Server,
-        mut conn: pal_async::pipe::PolledPipe,
-        transport: ResolvedTransport,
-    ) -> anyhow::Result<()> {
-        let mut first_byte = [0];
-        if conn.read(&mut first_byte).await? == 0 {
-            return Ok(());
-        }
-        let conn = PrefixedStream {
-            prefix: Some(first_byte[0]),
-            inner: conn,
-        };
-        serve_by_protocol(server, conn, first_byte[0], transport).await
-    }
-
-    #[cfg(windows)]
-    struct PrefixedStream<T> {
-        prefix: Option<u8>,
-        inner: T,
-    }
-
-    #[cfg(windows)]
-    impl<T: AsyncRead + Unpin> AsyncRead for PrefixedStream<T> {
-        fn poll_read(
-            mut self: std::pin::Pin<&mut Self>,
-            cx: &mut std::task::Context<'_>,
-            buf: &mut [u8],
-        ) -> std::task::Poll<std::io::Result<usize>> {
-            if buf.is_empty() {
-                return std::task::Poll::Ready(Ok(0));
-            }
-            if let Some(prefix) = self.prefix.take() {
-                buf[0] = prefix;
-                return std::task::Poll::Ready(Ok(1));
-            }
-            std::pin::Pin::new(&mut self.inner).poll_read(cx, buf)
-        }
-    }
-
-    #[cfg(windows)]
-    impl<T: AsyncWrite + Unpin> AsyncWrite for PrefixedStream<T> {
-        fn poll_write(
-            mut self: std::pin::Pin<&mut Self>,
-            cx: &mut std::task::Context<'_>,
-            buf: &[u8],
-        ) -> std::task::Poll<std::io::Result<usize>> {
-            std::pin::Pin::new(&mut self.inner).poll_write(cx, buf)
-        }
-
-        fn poll_flush(
-            mut self: std::pin::Pin<&mut Self>,
-            cx: &mut std::task::Context<'_>,
-        ) -> std::task::Poll<std::io::Result<()>> {
-            std::pin::Pin::new(&mut self.inner).poll_flush(cx)
-        }
-
-        fn poll_close(
-            mut self: std::pin::Pin<&mut Self>,
-            cx: &mut std::task::Context<'_>,
-        ) -> std::task::Poll<std::io::Result<()>> {
-            std::pin::Pin::new(&mut self.inner).poll_close(cx)
-        }
+        super::platform::serve_socket(server, conn, first_byte, transport, registry).await
     }
 
     /// Waits for the first byte of the stream to be available and returns it
@@ -486,39 +355,7 @@ pub struct TtrpcWorker {
     transport: ResolvedTransport,
 }
 
-enum BoundListener {
-    Unix(UnixListener),
-    #[cfg(windows)]
-    Pipe(pal_async::windows::pipe::NamedPipeServer),
-}
-
 pub const TTRPC_WORKER: WorkerId<Parameters> = WorkerId::new("TtrpcWorker");
-
-#[cfg(windows)]
-fn rpc_pipe_security_descriptor(
-    allow_sid: &str,
-) -> anyhow::Result<pal::windows::security::LocalSecurityDescriptor> {
-    use pal::windows::security::LocalSid;
-    use pal::windows::security::current_process_user_sid;
-    use windows_sys::Win32::Storage::FileSystem::FILE_APPEND_DATA;
-    use windows_sys::Win32::Storage::FileSystem::FILE_GENERIC_READ;
-    use windows_sys::Win32::Storage::FileSystem::FILE_GENERIC_WRITE;
-
-    anyhow::ensure!(
-        !allow_sid.contains('\0'),
-        "invalid RPC 'allow-sid': contains NUL"
-    );
-    let client: LocalSid = allow_sid.parse().context("invalid RPC 'allow-sid'")?;
-    let server = current_process_user_sid().context("failed to query RPC pipe owner")?;
-    let client_access = (FILE_GENERIC_READ | FILE_GENERIC_WRITE) & !FILE_APPEND_DATA;
-    format!(
-        "O:{server}D:P(A;;FA;;;{server})(A;;0x{client_access:x};;;{client})",
-        server = server.to_string_sid(),
-        client = client.to_string_sid(),
-    )
-    .parse()
-    .context("failed to build RPC pipe security descriptor")
-}
 
 impl Worker for TtrpcWorker {
     type Parameters = Parameters;
@@ -527,23 +364,7 @@ impl Worker for TtrpcWorker {
 
     fn new(parameters: Self::Parameters) -> anyhow::Result<Self> {
         Ok(Self {
-            listener: match parameters.listener {
-                Listener::Unix(listener) => BoundListener::Unix(listener),
-                #[cfg(windows)]
-                Listener::Pipe { path, allow_sid } => {
-                    let security_descriptor = allow_sid
-                        .as_deref()
-                        .map(rpc_pipe_security_descriptor)
-                        .transpose()?;
-                    BoundListener::Pipe(
-                        pal_async::windows::pipe::NamedPipeServer::create_with_security(
-                            path,
-                            security_descriptor.as_deref(),
-                        )
-                        .context("failed to create named pipe")?,
-                    )
-                }
-            },
+            listener: BoundListener::new(parameters.listener)?,
             transport: match parameters.transport {
                 #[cfg(feature = "ttrpc")]
                 RpcTransport::Ttrpc => ResolvedTransport::Ttrpc,
@@ -597,16 +418,9 @@ impl VmService {
         let server_task = self.driver.spawn("ttrpc-server", {
             let driver = self.driver.clone();
             async move {
-                let r = match listener {
-                    BoundListener::Unix(listener) => {
-                        dispatch::run(&server, &driver, listener, cancel_recv, transport, registry)
-                            .await
-                    }
-                    #[cfg(windows)]
-                    BoundListener::Pipe(listener) => {
-                        dispatch::run_pipe(&server, &driver, listener, cancel_recv, transport).await
-                    }
-                };
+                let r = listener
+                    .run(&server, &driver, cancel_recv, transport, registry)
+                    .await;
                 match &r {
                     Ok(()) => tracing::debug!("ttrpc server shutting down"),
                     Err(err) => tracing::error!(
@@ -2498,104 +2312,6 @@ fn build_vhost_user_device(
     _vhost_user: vmservice::VhostUser,
 ) -> anyhow::Result<Resource<VirtioDeviceHandle>> {
     anyhow::bail!("vhost-user is only supported on unix hosts")
-}
-
-#[cfg(all(test, windows))]
-mod named_pipe_tests {
-    use super::*;
-    use pal::windows::security::LocalSecurityDescriptor;
-    use pal::windows::security::current_process_user_sid;
-    use pal_async::windows::pipe::NamedPipeServer;
-    use std::fs::OpenOptions;
-    use std::os::windows::io::AsHandle;
-    use test_with_tracing::test;
-    use windows_sys::Win32::Security::DACL_SECURITY_INFORMATION;
-    use windows_sys::Win32::Security::OWNER_SECURITY_INFORMATION;
-
-    fn pipe_path() -> String {
-        let mut random = [0; 16];
-        getrandom::fill(&mut random).unwrap();
-        format!(
-            r"\\.\pipe\openvmm-rpc-test-{:032x}",
-            u128::from_ne_bytes(random)
-        )
-    }
-
-    fn assert_pipe_security(file: &File, expected: &str) {
-        let descriptor = LocalSecurityDescriptor::from_handle(
-            file.as_handle(),
-            OWNER_SECURITY_INFORMATION | DACL_SECURITY_INFORMATION,
-        )
-        .unwrap();
-        assert_eq!(descriptor.to_sddl().unwrap(), expected);
-    }
-
-    #[test]
-    fn reject_invalid_pipe_sid() {
-        for sid in [
-            "",
-            "not-a-sid",
-            "S-1-5-19)(A;;FA;;;WD)",
-            "S-1-5-19\0",
-            "S-1-5-19\0\0",
-            "S-1-5-19\0trailing",
-        ] {
-            assert!(
-                rpc_pipe_security_descriptor(sid).is_err(),
-                "accepted invalid SID: {sid:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn worker_pipe_authorizes_configured_sid() {
-        DefaultPool::run_with(async |driver| {
-            let path = pipe_path();
-            let worker = TtrpcWorker::new(Parameters {
-                listener: listener_from_path(path.as_ref(), Some("S-1-5-19")).unwrap(),
-                transport: RpcTransport::Auto,
-            })
-            .unwrap();
-            let BoundListener::Pipe(listener) = &worker.listener else {
-                panic!("expected named pipe listener");
-            };
-            let user = current_process_user_sid().unwrap();
-            let expected: LocalSecurityDescriptor = format!(
-                "O:{user}D:P(A;;FA;;;{user})(A;;0x12019b;;;S-1-5-19)",
-                user = user.to_string_sid()
-            )
-            .parse()
-            .unwrap();
-            let expected = expected.to_sddl().unwrap();
-            for _ in 0..2 {
-                let accept = listener.accept(&driver).unwrap();
-                let client = OpenOptions::new()
-                    .read(true)
-                    .write(true)
-                    .open(&path)
-                    .unwrap();
-                let server = accept.await.unwrap();
-                assert_pipe_security(&client, &expected);
-                assert_pipe_security(&server, &expected);
-            }
-        });
-    }
-
-    #[test]
-    fn worker_construction_reserves_pipe() {
-        let path = pipe_path();
-        let parameters = || Parameters {
-            listener: listener_from_path(path.as_ref(), None).unwrap(),
-            transport: RpcTransport::Auto,
-        };
-
-        let worker = TtrpcWorker::new(parameters()).unwrap();
-        assert!(NamedPipeServer::create(&path).is_err());
-        assert!(TtrpcWorker::new(parameters()).is_err());
-
-        drop(worker);
-        NamedPipeServer::create(&path).unwrap();
-    }
 }
 
 #[cfg(all(test, target_os = "linux"))]
